Revert renames and get tests working locally
