+++ exercises/concept/amusement-park-improvements/.meta/config.json
@@ -4,8 +4,8 @@
   "contributors": ["kotp", "iHiD", "kayn1"],
   "language_versions": ">=2.6.6",
   "files": {
-    "solution": ["amusement_park_improvements.rb"],
-    "test": ["amusement_park_improvements_test.rb"],
+    "solution": ["attendee.rb"],
+    "test": ["attendee_test.rb"],
     "exemplar": [".meta/exemplar.rb"]
   }
 }
