+++ exercises/concept/amusement-park/attendee.rb
