+++ exercises/concept/amusement-park/attendee_test.rb
@@ -1,5 +1,5 @@
 require 'minitest/autorun'
-require_relative 'amusement_park'
+require_relative 'attendee'
 
 class AttendeeTest < Minitest::Test
   def test_new_instance
