+++ exercises/concept/boutique-inventory-improvements/boutique_inventory.rb
